+++ src/main/java/nextstep/exception/AuthorizationException.java
@@ -1,10 +1,11 @@
-package nextstep.support;
+package nextstep.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthorizationException extends RuntimeException {
+
     public AuthorizationException() {
     }
