+++ src/main/java/nextstep/exception/DuplicateEntityException.java
@@ -1,4 +1,5 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class DuplicateEntityException extends RuntimeException {
+
 }
