+++ src/main/java/nextstep/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
