+++ src/main/java/nextstep/member/MemberController.java
@@ -1,16 +1,19 @@
 package nextstep.member;
 
-import nextstep.auth.AuthorizationExtractor;
-import nextstep.support.AuthorizationException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.servlet.http.HttpServletRequest;
 import java.net.URI;
+import javax.servlet.http.HttpServletRequest;
+import nextstep.auth.AuthorizationExtractor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/members")
 public class MemberController {
+
     private MemberService memberService;
 
     public MemberController(MemberService memberService) {
