+++ src/main/java/nextstep/theme/ThemeService.java
@@ -1,12 +1,12 @@
 package nextstep.theme;
 
-import nextstep.support.NotExistEntityException;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
+import nextstep.exception.NotExistEntityException;
+import org.springframework.stereotype.Service;
 
 @Service
 public class ThemeService {
+
     private ThemeDao themeDao;
 
     public ThemeService(ThemeDao themeDao) {
